requirements.txt
# Google ADK and AI dependencies
google-genai
google-cloud-aiplatform

# Database
pymongo[srv]>=4.6.0
motor>=3.3.0  # Async MongoDB driver (optional, for future async support)

# Environment management
python-dotenv>=1.0.0

# Data validation
pydantic>=2.5.0

# Utilities
python-dateutil>=2.8.2

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.12.0
flake8>=6.1.0

